Data & Automation Agents/CSV data cleaning agent/requirements.txt
pandas
openai
anthropic
google-generativeai
groq
streamlit
python-dotenv
plotly
pydantic
instructor
tabulate
fastapi
uvicorn
